basic navigation when receiving push notification
